Added Navigation Between Fragments

+++ app/src/main/java/com/holotube/util/CardViewUtils.kt
@@ -5,9 +5,9 @@ import java.text.ParsePosition
 import java.util.*
 
 object CardViewUtils {
-    fun getTimeDifference(inputDate: String): Long {
+    fun getTimeDifference(inputDate: String): Double {
         val startTime = ISO8601Utils.parse(inputDate, ParsePosition(0))
         val currentTime = Calendar.getInstance().time
-        return (startTime.time - currentTime.time) / (1000 * 60 * 60)
+        return (startTime.time.toDouble() - currentTime.time.toDouble()) / (1000 * 60 * 60)
     }
 }
